Refactor sidebar
